SI-4694 Extended the param_t structure with a 64-bit vaddr field to be used in case of mapping the parameter to VMEM storage. Changed the len/size from uint64_t to uint32_t since all memory copy operations in the VMEM domain can only be uint32_t since they involve a physical address.

+++ include/vmem/vmem.h
@@ -33,8 +33,8 @@ enum vmem_types{
 
 typedef struct vmem_s {
 	int type;
-	void (*read)(struct vmem_s * vmem, uint64_t addr, void * dataout, uint64_t len);
-	void (*write)(struct vmem_s * vmem, uint64_t addr, const void * datain, uint64_t len);
+	void (*read)(struct vmem_s * vmem, uint64_t addr, void * dataout, uint32_t len);
+	void (*write)(struct vmem_s * vmem, uint64_t addr, const void * datain, uint32_t len);
 	int (*backup)(struct vmem_s * vmem);
 	int (*restore)(struct vmem_s * vmem);
 	int (*flush)(struct vmem_s * vmem);
@@ -46,10 +46,10 @@ typedef struct vmem_s {
 	void * driver;
 } vmem_t;
 
-void * vmem_memcpy(void * to, const void * from, uint64_t size);
-void * vmem_write(uint64_t to, const void * from, uint64_t size);
-void * vmem_read(void * to, uint64_t from, uint64_t size);
-void * vmem_cpy(uint64_t to, uint64_t from, uint64_t size);
+void * vmem_memcpy(void * to, const void * from, uint32_t size);
+void * vmem_write(uint64_t to, const void * from, uint32_t size);
+void * vmem_read(void * to, uint64_t from, uint32_t size);
+void * vmem_cpy(uint64_t to, uint64_t from, uint32_t size);
 
 vmem_t * vmem_index_to_ptr(int idx);
 int vmem_ptr_to_index(vmem_t * vmem);
